% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.r
\name{get_climate_status}
\alias{get_climate_status}
\title{Retrieve current status of global effort to combat climate change}
\usage{
get_climate_status(temp, investment, support)
}
\arguments{
\item{temp}{- Average global temperature in degrees F}

\item{investment}{- Total global investment in climate change
scale of ("minimal", "decent", "fully vested")}

\item{support}{- global population support to combat climate change ("low", "medium", "high")}
}
\value{
- Message on the status of combating climate change
}
\description{
Retrieve current status of global effort to combat climate change
}
\examples{
  get_climate_status(88, "minimal", "low")
}
